### 2024-05-24 #6777 @editorialbot
[PRE REVIEW]: Desalination and brine treatment systems integrated modelling framework: simulation and evaluation of water and resource recovery
**Submitting author:** @rodoulak (<a href="http://orcid.org/0000-0002-9812-8963">Rodoula Ktori</a>)
**Repository:** https://github.com/rodoulak/Desalination-and-Brine-Treatment-Simulation-.git
**Branch with paper.md** (empty if default branch): main
**Version:** DesalSim 1.0.1 (https://pypi.org/project/DesalSim/)
