### 2024-05-28 #6777 @kthyng
@rodoulak can you work on troubleshooting the paper compilation errors? You can recompile with the command I used.
