+++ paper/paper.bib
@@ -62,7 +62,6 @@
   year={2002},
   publisher={Elsevier}, 
   doi={10.1016/B978-0-444-50810-2.50018-X}
-  url={https://doi.org/10.1016/B978-0-444-50810-2.50018-X}
 }
 @article{kesieme2013economic,
   title={Economic analysis of desalination technologies in the context of carbon pricing, and opportunities for membrane distillation},
