### 2024-05-31 #6777 @rodoulak
@kthyng I checked the error, and there was a problem with line 65 in the bib file. I deleted that line (URL). I don't know how to generate the PDF, so I can test if it works now.
