+++ DesalSim.egg-info/requires.txt
@@ -1,7 +1,7 @@
-numpy==1.26.4
-pandas==2.2.2
-scipy==1.13.0
-matplotlib==3.8.4
-scikit-learn==1.4.2
-plotly==5.22.0
-openpyxl==3.1.2
+numpy
+pandas
+scipy
+matplotlib
+scikit-learn
+plotly
+openpyxl
