### 2024-06-13 871f5e8
Update TCr_Tutorial.md
