### 2024-06-20 #6777 @AdamRJensen
> @editorialbot invite @AdamRJensen as editor

Unfortunately not, I'm running busy with three other JOSS submissions and other work
