### 2024-07-02 #6777 @elbeejay
@rodoulak - would you consider cleaning up the repository a bit before we send it out for review? It looks like a lot of Python auxiliary files and folders that are typically skipped over by git (such as the build/lib, dist, .egg-info folders) are in your repository.

I'd suggest looking at [GitHub's Python `.gitignore` file](https://github.com/github/gitignore/blob/main/Python.gitignore) and potentially adopting it to help manage the files that get pushed up to the GitHub repository. I am happy to hear reasons for why all of those files would need to be included, but it is atypical of a Python project.

Thanks,
Jay
