+++ build/lib/Desalsim/ro_unit_f.py
@@ -1,104 +1,0 @@
-import math
-from desalsim import constants 
-from desalsim.density_calc import density_calc 
-#%%
-#%%constants
-R=8.31446261815324 #gas constant
-T=20+273.15 #K
-MW_Na=22.99
-MW_cl=35.453
-MW_so4=32.064+4*15.999
-MW_K=39.102
-MW_Ca=40.08
-MW_Mg=24.31
-MW_HCO3=1.008+12.011+3*15.999
-#assumptions 
-dp=2
-
-#molarity 
-class molarity:
-    def __init__(self, MW, zi, Ci):
-        self.MW=MW
-        self.zi=zi
-        self.Ci=Ci
-        self.meq=self.Ci*1000*self.zi/self.MW
-        #self.mi=self.Ci/(self.MW*(1000-self.sum_conc)/1000                    #CHECK AGAIN THE UNITS 
-        
-#%%
-class nfmass:
-    def __init__(self, comp, Cfeedi, rjr, Wrec, Qf):
-        self.comp=comp
-        self.Cfeedi=Cfeedi #ion concentration g/l
-        self.rjr=rjr #rejection of each ion of the membrane 
-        self.Wrec = Wrec # water recovery first pass 
-        self.Qf=Qf
-        
-                       
-    def permcal(self):
-        self.Qperm= self.Wrec*self.Qf
-        self.Qconc=self.Qf-self.Qperm
-        self.Cpermi = (1-self.rjr)*self.Cfeedi
-        self.Cconci = (self.Qf*self.Cfeedi-self.Qperm*self.Cpermi)/self.Qconc
-#        print(" ion concenctration of " + self.comp+ " in permeate stream is"+" "+ str(self.Cpermi))
-#        print(" ion concenctration of " + self.comp+ " in concentrate stream is"+" "+str(self.Cconci))
-
-#%% osmotic pressure 
-class Osmotic_pressure:
-    def __init__(self, C1, z1,  C2, z2,  C3, z3,  C4, z4, C5, z5, C6, z6):
-        self.C1=C1
-        self.z1=z1
-        self.C2=C2
-        self.z2=z2
-        self.C3=C3
-        self.z3=z3
-        self.C4=C4
-        self.z4=z4
-        self.C5=C5
-        self.z5=z5
-        self.C6=C6
-        self.z6=z6
-        self.sum_Ci=sum([C1, C2, C3, C4, C5, C6])
-        self.zi_2=[z1**2, z2**2, z3**2, z4**2, z5**2, z6**2]
-        self.conc=[C1/MW_Na, C2/MW_cl, C3/MW_K, C4/MW_Mg, C5/MW_Ca, C6/MW_so4] #mol/l
-        self.mi=[C1*1000/(MW_Na*1000*((1e+6-self.sum_Ci*1000)/1e+6)), C2*1000/(MW_cl*1000*((1e+6-self.sum_Ci*1000)/1e+6)), C3*1000/(MW_K*1000*((1e+6-self.sum_Ci*1000)/1e+6)), C4*1000/(MW_Mg*1000*((1e+6-self.sum_Ci*1000)/1e+6)), C5*1000/(MW_Ca*1000*((1e+6-self.sum_Ci*1000)/1e+6)), C6*1000/(MW_so4*1000*((1e+6-self.sum_Ci*1000)/1e+6))]
-        self.mizi_2=[]
-        
-    # def osmotic_pressure_calculation(self):
-        for i in range(6) :
-            self.mizi_2.append(self.mi[i]*self.zi_2[i])
-        self.SI=sum(self.mizi_2)/2
-        self.B=-348.662/(T)+6.72817-0.971307*math.log(T)
-        self.C=40.5016/(T)-0.721404+0.103915*math.log(T)
[…]
