### 2024-07-16 #6777 @rkingsbury
> @rkingsbury would you be willing to review this submission to JOSS titled: _Desalination and brine treatment systems integrated modelling framework: simulation and evaluation of water and resource recovery_? I know you previously published your Python water chemistry package, `pyEQL`, with JOSS, and thought that you might be able to provide some good feedback for this submission, especially on the desalination and brine treatment modeling.
>
> Please let us know if you'd be willing to provide a review. I understand if you do not have the time to commit to a review for JOSS at this time.

Hi @elbeejay I am interested to review this and I think I'm well qualified to do so. However, I do have limited bandwidth right now so my review might be slow. What is the expected timeframe?
