### 2024-07-20 #6777 @elbeejay
@stroutm would you be willing to review this submission to JOSS titled "Desalination and brine treatment systems integrated modelling framework: simulation and evaluation of water and resource recovery"? I understand you may already be familiar with JOSS from your publication on the `rrcf` package a few years ago. At a high-level, this submission deals with the simulation and modeling of an integrated water system, which I thought might be a good fit for someone with your expertise.

Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time. If you cannot serve as a reviewer at this time but have a peer, collaborator, student, or colleague who might be available and would be a good fit for this submission, please let me know!

Thanks,
Jay
