### 2024-07-22 #6777 @DCoppitters
> @rkingsbury would you be willing to review this submission to JOSS titled: _Desalination and brine treatment systems integrated modelling framework: simulation and evaluation of water and resource recovery_? I know you previously published your Python water chemistry package, `pyEQL`, with JOSS, and thought that you might be able to provide some good feedback for this submission, especially on the desalination and brine treatment modeling.
>
> Please let us know if you'd be willing to provide a review. I understand if you do not have the time to commit to a review for JOSS at this time.

Hi @elbeejay,

Thank you for the invitation. Unfortunately, my schedule doesn't allow for additional reviews at the moment. I'm very sorry about that.
