### 2024-07-22 #6777 @DCoppitters
> @DCoppitters would you be willing to review this submission to JOSS titled: _Desalination and brine treatment systems integrated modelling framework: simulation and evaluation of water and resource recovery_? I understand that desalination and brine treatment systems might not be your niche, but the submission is presenting more than that, it is an integrated framework consisting of both the water chemistry and the associated economic models. This seems, at a high-level, similar in concept to your RHEIA software, which was previously published in JOSS and allows for the optimization of multiple objectives when designing hydrogen-energy systems.
>
> Please let us know if you'd be willing to provide a review. I understand if you do not have the time to commit to a review for JOSS at this time.

Hi @elbeejay,

Thank you for the invitation. Unfortunately, my schedule doesn't allow for additional reviews at the moment. I'm very sorry about that.
