### 2024-07-22 #6777 @elbeejay
@jennwuu and @bemcdonnell - would either of you be interested in reviewing this submission to JOSS titled "Desalination and brine treatment systems integrated modelling framework: simulation and evaluation of water and resource recovery"? JOSS is an open-source, diamond access software journal. I'm reaching out as you are both co-authors on the JOSS publication of `PySWMM`, which I recognize is different from this submission, but I believe much of your expertise is transferrable and would make you valuable reviewers for this submission. Below I've included a summary of the JOSS review process as it has been a few years since the PySWMM paper came out.

At JOSS we do open checklist-driven reviews; peer-review criteria can be viewed [here](https://joss.readthedocs.io/en/latest/review_criteria.html). This issue is a "pre-review issue" which we use to find peer-reviewers. Once 3 reviewers are found, we will officially start the review in a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks, although this can be extended if needed. If you are not able to review but can recommend someone else, please mention them here (in this case please mention their GitHub handle without the "@" symbol).

If you are interested, please take a look at the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) to ensure you do not have a conflict before agreeing to review this submission.

Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time. If you cannot serve as a reviewer at this time but have a peer, collaborator, student, or colleague who might be available and would be a good fit for this submission, please let me know!

Thanks,
Jay
