### 2024-07-30 #7062 @elbeejay
@rkingsbury, @yalinli2, @BenWinchester thanks for agreeing to review for JOSS. Instructions for generating your reviewer checklist and conducting your review are in the first comment of this issue thread, but feel free to reach out if you have any questions.

The review process at JOSS is quite fluid, you don't need to do a whole review at once, you are free to tackle this piece-by-piece if you choose, and converse with the author as needed.

Similarly, we ask reviewers to try and complete their reviews within 6 weeks (so by Sept 10 in this case), but we can also be flexible and accommodate extensions. All I ask is that you just communicate your plan clearly with us here so the authors and the other reviewers here understand what is going on. Especially with summer holidays I understand that folks may be away, so please just keep us informed, thank you!
