### 2024-08-21 #7062 @elbeejay
Hello @yalinli2, @BenWinchester, and @rkingsbury - we are about 3 weeks in, so halfway, through the standard review time period. Just trying to ping you all so this doesn't slip off of to-do lists.

Again, please let us know here if you need an extension for whatever reason.

We appreciate open-communications as we attempt to keep the JOSS review process as transparent and friendly (to all involved) as possible.
