### 2024-09-05 87629ba
update examples and tutorials
