### 2024-09-05 #7062 @rodoulak
> ## Review checklist for @yalinli2
> ### Conflict of interest
> * [x]  I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [x]  **Repository:** Is the source code for this software available at the https://github.com/rodoulak/Desalination-and-Brine-Treatment-Simulation-.git?
>
> Source codes available.
>
> * [x]  **License:** Does the repository contain a plain-text LICENSE or COPYING file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
>
> MIT license.
>
> * [x]  **Contribution and authorship:** Has the submitting author (@rodoulak) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> All commits were from rodoulak. The paper includes additional contributors.
>
> * [x]  **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
> The package `desalsim` contains simulation models for desalination and brine treatment technologies including nanofiltration/reverse osmosis, multi-effect distillation, thermal crystallization, chemical precipitation, eutectic freeze crystallization, and electrodialysis with membrane.
>
> Though the mathematical development of these models have been described elsewhere, the Python implementation would be useful to researchers.
>
> * [x]  **Data sharing:** If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
>
> The paper contains no original data.
>
> * [x]  **Reproducibility:** If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
>
> The paper contains no original results (there are tutorials, examples, and tests included in the package, see comments below on them).
>
> * [x]  **Human and animal research:** If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item.
>
> The paper contains no data on humans subjects or animals.
>
> ### Functionality
> * [x]  **Installation:** Does installation proceed as outlined in the documentation?
>
> `pip install desalsim` worked, but `pip install -r requirements.txt` didn't, `requirements.txt` file missing.
>
> * [x]  **Functionality:** Have the functional claims of the software been confirmed?
>
> The examples and tutorials largely worked, but some of them need to be updated. E.g., `RuntimeWarning` triggered for the default values in example 1: ![image](https://private-user-images.githubusercontent.com/49011373/363493584-d8dbe48f-a775-4934-9b7c-076145d64e67.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjU1NTYyMzIsIm5iZiI6MTcyNTU1NTkzMiwicGF0aCI6Ii80OTAxMTM3My8zNjM0OTM1ODQtZDhkYmU0OGYtYTc3NS00OTM0LTliN2MtMDc2MTQ1ZDY0ZTY3LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA5MDUlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwOTA1VDE3MDUzMlomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWYyZjFhOGNjNjQxZTQ3MDA0MDk0YjFhMTUzYTAxZWRkODU5YmU4ZjdjOGFlYjNiNGM5Y2MwOTFlZjc2OTJjZWYmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.1opzVS6Kd0cL_jTI1De-4Ia26ngIBbYLWyOzAYFlRPM)
>
> In some tutorials, `Desalsim` is used instead of `desalsim`, led to some importing errors.
>
> Other minor errors in the tutorials that prevent new users to directly run them, e.g.,
>
> `Qf_nf` was not defined in
>
> https://github.com/rodoulak/desalsim/blob/8a2dc8dbf7d46574e23c3cf97259fa61c2d483e7/example/Example_1_Tutorial.md?plain=1#L118
>
> Should this be just `Qf`?
>
> I'm not sure if there are packages that can test `.md` files, but if the tutorials are in Jupyter Notebook, we can use automated tests.
>
> * [x]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> There are no performance claims of the software.
>
> ### Documentation
> * [x]  **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>
> Motivation/purpose statement included in README and the paper.
>
> * [x]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
>
> Installation instructions are provided, though in some places `Desalsim` should be replaced with `desalsim`.
>
> * [x]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
> Examples included, but the authors are recommended to double-check the tutorials to make sure they've been updated throughout. Would be good to add automated tests.
>
> * [x]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>
> A separate [readthedocs](https://desalsim-web.readthedocs.io/en/latest/) page is available.
>
> * [x]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?
>
> Test files available, but recommend renaming the scripts (e.g., start with `test_`) so that they can be auto-detected by `pytest`. Also, one of tests failed: ![image](https://private-user-images.githubusercontent.com/49011373/363494485-e8d33cd2-f748-492c-8609-5f8900bf19c9.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjU1NTYyMzIsIm5iZiI6MTcyNTU1NTkzMiwicGF0aCI6Ii80OTAxMTM3My8zNjM0OTQ0ODUtZThkMzNjZDItZjc0OC00OTJjLTg2MDktNWY4OTAwYmYxOWM5LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA5MDUlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwOTA1VDE3MDUzMlomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWRiY2FiNzZhMDNlMzRlZjk1ODJjNGUyMzAxNmExY2ExZjIxY2RhNGFkMDlhYmZhODFhNGZhYmQwZDBmZTA0NTEmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.oQ9LksaKcpUdkaGSVTgmYJwpWbOMVW1RkvZIymJNZjQ)
>
> * [x]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> Short description on how to contribute included in README.
>
> ### Software paper
> * [x]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> * [x]  **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
> * [x]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
>
> The paper indicated a general lack of OSS for simulation of desalination and minerals recovery systems. No comparison was provided.
>
> * [x]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
>
> The paper was well written. Only a possible misspelling (`valisated` should probably be `validated`). Also maybe want to capitalize the "resource" in the tags.
>
> There are some issues in the [mathematical description](https://github.com/rodoulak/desalsim/blob/main/paper/Mathematical_description.pdf):
>
> * No Table B.1 was provided.
> * Extra blank space in the first paragraph of A.6. for EDBM model.
>
> Please note that I didn't check the actual equations - I assume they and their implementation in the package are all correct. I'm not a desal person, so I'm not aware if there are good benchmark models that can be used to test the accurate of certain models. Maybe other reviewers have insight.
>
> * [x]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?
>
> A separate bibliography file is included, looks good as in the proof.
>
> ### Comments to be addressed:
> * [x]  Add requirements.txt.
> * [x]  Fix minor bugs in tutorials/examples.
> * [x ]  Automate test, can consider adding a GitHub workflow.
> * [x]  Fix the issues in mathematical description.
