### 2024-09-06 #7062 @yalinli2
Hi @rodoulak , just one minor thing in the test, I think the `expected_solid_mass` here is off by 1000, probably shouldn't include the 1000 factor in the calculation, thanks!
https://github.com/rodoulak/desalsim/blob/e6389ba87bfa2ed9b9152c4dc1bc7ce2fb579b86/tests/test_thermal_cryst.py#L38
