### 2024-09-09 #7062 @rodoulak
Hi @yalinli2,
Thank you for your comment.

I updated the line where the solution density was calculated because the units were incorrect. Previously, the density was in kg/m³ instead of kg/L, which is why the [tests/test_thermal_cryst.py](https://github.com/rodoulak/desalsim/blob/main/tests/test_thermal_cryst.py) tests were off by a factor of 1000.

Here is the revised line: [tests/test_thermal_cryst.py#L21](https://github.com/rodoulak/desalsim/blob/main/tests/test_thermal_cryst.py#L21).

Thanks again for your attention to detail!
