### 2024-09-09 #7062 @rkingsbury
> @elbeejay thanks for the reminder. I am also optimistic that I can finish within the next 3 weeks; planning to get started toward the end of this week. At worst I'd anticipate I might as for 1 additional week, but if I think I'll need to do so, I'll let you know here.

Hi @elbeejay just to follow up on the above - I think I will need until at least the end of this week, possibly into next week, to finish this review. Apologies that I haven't been able to move forward more quickly, but I haven't forgotten!
