### 2024-09-09 #7062 @rkingsbury
@rodoulak I have begun my review and left some comments concerning the "Software Paper" section. Please see above and let me know your thoughts. I haven't dug much into the software itself yet, but will do so soon.
