### 2024-09-10 #7062 @BenWinchester
@elbeejay , sorry for the delay in getting back to you! I've now gone through the checklist and have comments below :smiley:. In general, the code is in good shape, but there are quite a few errors in the examples which, ideally, should all be fixed before the software comes back for a second round of reviewing: I've done my best to highlight errors in the example scripts, but I likely haven't been exhaustive (and was unable to complete the examples as the number of undefined variables meant that I could no longer guess what variables should be implemented) and have missed a few, so I recommend that the authors go through the documentation thoroughly and check for any and all errors :smile:.

  I encountered a few errors in the examples:
  - In [Tutorial 1](https://github.com/rodoulak/desalsim/blob/main/example/Example_1_Tutorial.md#311-define-feed-characteristics), there is an error in the code snipped in Section 3.1.1.:
    ```python
    Qsw = 3000 / 24 * d_in #kg/hr
    ```
    where `d_in` has not been defined yet as it is only instantiated in the next code snippet;
  - A code snippet [further down the page](https://github.com/rodoulak/desalsim/blob/main/example/Example_1_Tutorial.md#calculate-energy-consumption) does not work, as a positional argument follows a keyword argument, which is not permitted in Python:
    ```python
    nf_energy=NfEnergy(P_osmo_c, P_osmo_f, P_osmo_p, dp=2, d_p, Qperm, Qf, d_in,n=0.8)
    ```
    and so the code fails. One fix may be
    ```diff
    - nf_energy=NfEnergy(P_osmo_c, P_osmo_f, P_osmo_p, dp=2, d_p, Qperm, Qf, d_in,n=0.8)
    + nf_energy=NfEnergy(P_osmo_c, P_osmo_f, P_osmo_p, (dp:=2), d_p, Qperm, Qf, d_in,n=0.8)
    ```
    if using Python 3.8 or newer;
  - The same line of code fails as the variable `d_p` hasn't been defined yet in the example code;
  - [Further down the same tutorial](https://github.com/rodoulak/desalsim/blob/main/example/Example_1_Tutorial.md#create-mfpfrcalc-objects), the variable `MFPFRCALC` has yet to be imported into the example:
    ```python
    # Create an instance of the inputpar class with the defined parameters
    mfpfr_dat = MFPFRCALC(Qin_mfpfr, Cin_mfpfr, *C_NaOH, *conv)
    ````
    and so the code tutorial fails at this line also;
  - [Further down still](https://github.com/rodoulak/desalsim/blob/main/example/Example_1_Tutorial.md#calculate-outlet-concentration) in the same `Example_1_Tutorial.md` file, the code relies on objects which exist within the `Desalsim.constants` module but which are not imported in the examples:
    ```python
    MW = [MW_Na, MW_Ca, MW_Cl, MW_K, MW_Mg, MW_SO4]
    ```
    and so, without these imports correctly being handled, the example tutorial fails at this line also;
  - Further, in the above smippet, the variables `MW_Cl` and `MW_SO4` do not exist in the `constants` module. However, `MW_cl` and `MW_so4` do exist, with lowercase letters, suggesting a typo in addition to the missing import highlighted. **Note:** this same typo exists in [later snippet(s)](https://github.com/rodoulak/desalsim/blob/main/example/Example_1_Tutorial.md#calculate-amount-of-hcl-for-ph-neutralization-and-the-final-outlet-concentration-from-mf-pfr-unit-after-ph-neutralization) as well;
  - `dp` being undefined also affects the [Calculate energy consumption](https://github.com/rodoulak/desalsim/blob/main/example/Example_1_Tutorial.md#calculate-energy-consumption-1) section of the example;
  - The class defined in the example to use to [summarise the code](https://github.com/rodoulak/desalsim/blob/main/example/Example_1_Tutorial.md#41-summarise-results) cannot be instantiated as described in the example:
    ```python
    tec1=indic("NF", Qconc, Qf, Qperm, "none", 0, "none", E_el_nf, 0, Ci_in, Cconc, QHCl_nf, Qantsc_nf)
    tec1.techn_indi()
    ```
    relies on the variable `QHCl_nf` which isn't instantiated elsewhere in the example. Further, this isn't a variable defined in the unimported `constants` module. (As an aside, I'm not a fan, in general, of defining classes and functions that don't have well-documented APIs and this class in the examples section is simply defined without any documentation or dosctrings to explain how the class functions or summarises the outputs. If this class is something that a user would be expected to implement and utilise on their own, _i.e._, if it has more of a use than simply summarising the outputs from the tutorial, then it would be nice to see a dosctring in there.);
  - I'm not convinced that the `__init__.py` file will work as the authors intend. The absolute naming of the imports used in the file,
    ```python
    # __init__.py
    from desalsim.density_calc import density_calc
    ... etc ...
    ```
    appears to import from the installed package, `desalsim`, rather than from the directory. So, when installing the package from the Github download using `python -m pip install -e .`, an error occurs in this file:
    ```python
    >>> import Desalsim
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
      File "C:\Users\bened\PhD Synced\Reviewing\desalsim\Desalsim\__init__.py", line 2, in <module>
        from desalsim.density_calc import density_calc
    ModuleNotFoundError: No module named 'desalsim'
    ```
    which appears to be because installing the code in this way embeds code which relies on the package being installed from pypi as well.
  - In the [Economic tutorial](https://github.com/rodoulak/desalsim/blob/main/example/Economic_Tutorial.md), I encountered errors as well. There's an error in the code snippet, where `constants` is imported. This import only works if the user runs Python from within the `examples` directory, which isn't clear within the example;
  - [Later down](https://github.com/rodoulak/desalsim/blob/main/example/Economic_Tutorial.md#24-import-results-from-technical-model) the example, the code imports the file `example_1.py` and relies on the import interpretation generating variables:
    > Then create lists with the results.

    However, these variables aren't accessible once the import code has finished running. Whilst they do exist within `example_1.py`, they are not accessible from a Python session as is suggested:
    ```python
    >>> import example_1
    Permeate stream flow rate is 89974.58kg/hr
    ...
    [1 rows x 15 columns]
    >>> s_conc=E_th_all
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
    NameError: name 'E_th_all' is not defined
    ```
    and so this example, too, is not runable as it fails at this point.
  **Note:** the example scripts, `example_1.py` _etc._, do run when executed with `python example_1.py`, but they do not work as described in the markdown documentation;
  In general, I feel that the API methods aren't satisfactoraly documented throughout the code. For instance, `density_calc` is the first function exposed as an external API in the `__init__.py` file, but the code lacks a function docstring:
  ```python
  def density_calc(T, S):
      A= 8.24493E-1 - 4.0899E-3*T + 7.6438E-5*T**2 -8.2467E-7*T**3 + 5.3675E-9*T**4
      B = -5.724E-3 + 1.0227E-4*T - 1.6546E-6*T**2
      C = 4.8314E-4
      rho = 1000*(1 -(T+288.9414)/(508929.2*(T+68.12963))*(T-3.9863)**2)
      rhos = rho + A*S + B*S**(3/2) + C*S**2
      return rhos
  ```
  and so it is unclear what T and S are (what type is required, even) and what a user can expect the API to return. I understand that a likely user of the package would be clear on the Physics/Engineering involved in desalination calculations, as it stands, the API is only understandable by learning through the examples and not by reading anything in the code.
  With only 15 externally-exposed APIs in `__init__.py`, there aren't a great deal of dosctrings to write for this package, but I feel that they would immensely help any user who is hoping to install the package and implement their own plant configuration.

I also encountered some deprecation warnings in the code, which it would be nice to have cleared up:
```bash
example\efc_unit_f.py:469: DeprecationWarning: Conversion of an array with ndim > 0 to a scalar is deprecated, and will error in future.
```
There were several caused by this example, on lines 432, 469, 478, , 486, 523, 532, and 540.

With regards the paper, I thought it was written well. The only comment I have is in response to authoship:
> Does the full list of paper authors seem appropriate and complete?

For authorship on JOSS papers, general membership of, or adminstration of, a research group isn't sufficient to be an author. As only the submitting author appears to have contributed to the repository, it's unclear what role(s) these other authors played. However, if the submitting author is happy that everyone listed on the paper was actively involved in the project, then I think this is fine; I just thought I'd note it here, but I don't have an issue with the list :smiley:
