+++ paper/paper.bib
@@ -129,12 +129,14 @@
   doi={10.1016/j.jece.2021.105338},
   url={https://doi.org/10.1016/j.jece.2021.105338}
 }
-@article{peters1968plant,
-  title={Plant design and economics for chemical engineers},
+@book{peters1968plant,
+  title={Plant Design and Economics for Chemical Engineers},
   author={Peters, Max Stone and Timmerhaus, Klaus D},
-  journal={(No Title)},
-  year={1968}
+  edition={5th},
+  publisher={McGraw-Hill Chemical Engineering Series},
+  year={2003},
 }
+
 @article{poirier2022techno,
   title={Techno-economic analysis of brine treatment by multi-crystallization separation process for zero liquid discharge},
   author={Poirier, Kristofer and Al Mhanna, Najah and Patchigolla, Kumar},
