### 2024-09-11 927ab3e
Update paper.bib
