+++ paper/paper.bib
@@ -74,15 +74,7 @@
   doi={10.1016/j.desal.2013.03.033},
   url={https://doi.org/10.1016/j.desal.2013.03.033}
 }
-@article{ktori2024sustainability,
-  title={Sustainability assessment framework for integrated desalination and resource recovery: a participatory approach},
-  author={Ktori, Rodoula and Parada, Mar Palmeros and Rodriguez-Pascual, Marcos and van Loosdrecht, Mark CM and Xevgeno, Dimitrios},
-  journal={Authorea Preprints},
-  year={2024},
-  publisher={Authorea}, 
-  doi={10.36227/techrxiv.170595103.36689587/v1},
-  url={http://dx.doi.org/10.36227/techrxiv.170595103.36689587/v1}
-}
+
 @article{micari2020techno,
   title={Techno-economic analysis of integrated processes for the treatment and valorisation of neutral coal mine effluents},
   author={Micari, Marina and Cipollina, A and Tamburini, A and Moser, M and Bertsch, Valentin and Micale, Giorgio},
