### 2024-09-11 cfbc595
Update paper.md
