+++ paper/paper.md
@@ -68,7 +68,7 @@ The proposed software is not intended to be a substitute for sophisticated physi
 # Acknowledgements 
 The software was developed by Rodoula Ktori, with theoretical support from all co-authors. Technological experts conducted the validation of each simulation model for the respective technology: Nanofiltration: Dionysia Diamantidou, Niels van Linden; Multi-effect Distillation: Alessandro Trezzi; MF-PFR: Fabrizio Vassallo, Carmelo Morgante, Andrea Cipollina; EDBM: Giovanni Virruso, Andrea Culcasi; EFC: Marcos Rodriguez Pascual.
 
-The technical process models were developed based on the report from [@xevgenos2023deliverable] and the following literature [@nayar2019cost, @morgante2022valorisation, @cassaro2023electrodialysis]. Then they were validated with experimental results from [@morgante2024pioneering]. The development of economic models were influenced by [@peters1968plant, @abraham2011socio, @bilton2011feasibility, @choi2015economic, @kesieme2013economic]. The analysis and comparison were developed based on [@ktori2024designing]. Detailed mathematical descriptions of the simulation equations and economic models are provided in the Supplementary Information, [Mathematical Description](paper/Mathematical_description.pdf).
+The technical process models were developed based on the report from [@xevgenos2023deliverable] and the following literature [@nayar2019cost, @morgante2022valorisation, @cassaro2023electrodialysis]. Then they were validated with experimental results from [@morgante2024pioneering]. The development of economic models were influenced by [@peters1968plant, @abraham2011socio, @bilton2011feasibility, @choi2015economic, @kesieme2013economic]. The analysis and comparison were developed based on [@ktori2024designing]. Detailed mathematical descriptions of the simulation equations and economic models are provided in the Supplementary Information, [Mathematical Description](https://github.com/rodoulak/Desalination-and-Brine-Treatment-Simulation-/tree/main/paper/Mathematical_description.pdf).
 
 This work was supported by the EU within the WATER MINING (Next generation water-smart management systems: large scale demonstrations for a circular economy and society) - Horizon 2020 research and innovation programme under grant agreement No 869474.
 
