### 2024-09-11 6edeb8a
Update paper.bib
