+++ paper/paper.bib
@@ -185,3 +185,9 @@ Micale, G.},
   year={2024},
   publisher={Authorea}
 }
+@misc{watertap,
+  author      = {WaterTAP contributors},
+  title       = {WaterTAP: An open-source water treatment model library (Version 0.6)},
+  year        = {2024},
+  url         = {https://github.com/watertap-org/watertap},
+}
