+++ paper/paper.md
@@ -57,13 +57,12 @@ Our software offers a comprehensive modelling framework for simulating desalinat
 Through our software's simulations, researchers, engineers, and policymakers gain the power to evaluate the resource recovery potential, economics, and greenhouse gas emissions associated with different configuration combinations. This empowerment with crucial information for early-stage design and strategic planning is a significant step toward fostering more sustainable water management practices.
 
 # Statement of need
-Traditionally, simulation models were developed to evaluate the influence of certain parameters on the characteristics of the recovered products and the performance of the technology in terms of energy, chemicals, and water consumption. In the desalination field, there is a lack of open-access simulation models. While commercial software programs, like WAVE [@dupontwebsite], exist for membranes, and numerous publications discuss techno-economic models for desalination [@el2002fundamentals] and brine treatment technologies [@xevgenos2015design, @micari2020techno, @chen2021zero, @panagopoulos2021beneficiation, @poirier2022techno, @morgante2022valorisation], there is a noticeable absence of open-access simulation tools in the literature. With the shift towards circular systems and integrated desalination and brine treatment technologies for resource recovery, there is a need for a unified tool that integrates simulation models for each technology. 
-Simulation and evaluation are essential for informing future design and decision-making processes. An open-access simulation tool is essential in the field of desalination to improve the credibility of evaluations, facilitate study repeatability, support validation efforts, and enable comparisons.
-Therefore, while the modelling of desalination and brine treatment processes has been extensively studied in the literature, our software offers an integrated platform for integrating these models and facilitating seamless data exchange. Based on this need, we provide, for the first time, open-source software that brings together a range of technical processes and economic models developed over the last decades. By making these models accessible and transparent, our software aims to advance research, engineering, and decision-making in the field of desalination and resource recovery.
-The proposed software provides a variety of examples to help modellers design and evaluate different technical configurations. 
+Traditionally, simulation models were developed to evaluate the influence of certain parameters on the characteristics of the recovered products and the performance of the technology in terms of energy, chemicals, and water consumption. However, in the desalination field, open-access simulation tools are notably lacking. While commercial software programs, like WAVE [@dupontwebsite], exist for membranes, and numerous publications discuss techno-economic models for desalination [@el2002fundamentals] and brine treatment technologies [@xevgenos2015design, @micari2020techno, @chen2021zero, @panagopoulos2021beneficiation, @poirier2022techno, @morgante2022valorisation], there is a noticeable absence of open-access simulation tools in the literature. The WaterTAP platform [@watertap] offers an open-source library for modelling water treatment technologies like reverse osmosis and electrodialysis. While it provides valuable simulation capabilities, it mainly focuses on desalination technologies and lacks extensive brine treatment.
+With the shift towards circular systems and integrated desalination and brine treatment technologies for resource recovery, there is a need for a unified tool. Our software addresses this need by integrating a diverse range of technologies—reverse osmosis, nanofiltration, multi-effect distillation, chemical precipitation, eutectic freeze crystallization, electrodialysis, and thermal crystallization—into a comprehensive platform. This platform not only models these processes but also provides detailed techno-economic and environmental analyses.
+The software provides a variety of examples to help modellers design and evaluate different technical configurations. An open-access simulation tool is crucial for enhancing the credibility, repeatability, and comparability of desalination studies, and for supporting informed design and decision-making. By offering transparent and accessible models, our software aims to advance research and practice in desalination and resource recovery.
 
 # Limitations 
[…]
