+++ paper/paper.md
@@ -58,7 +58,9 @@ Through our software's simulations, researchers, engineers, and policymakers gai
 
 # Statement of need
 Traditionally, simulation models were developed to evaluate the influence of certain parameters on the characteristics of the recovered products and the performance of the technology in terms of energy, chemicals, and water consumption. However, in the desalination field, open-access simulation tools are notably lacking. While commercial software programs, like WAVE [@dupontwebsite], exist for membranes, and numerous publications discuss techno-economic models for desalination [@el2002fundamentals] and brine treatment technologies [@xevgenos2015design, @micari2020techno, @chen2021zero, @panagopoulos2021beneficiation, @poirier2022techno, @morgante2022valorisation], there is a noticeable absence of open-access simulation tools in the literature. The WaterTAP platform [@watertap] offers an open-source library for modelling water treatment technologies like reverse osmosis and electrodialysis. While it provides valuable simulation capabilities, it mainly focuses on desalination technologies and lacks extensive brine treatment.
+
 With the shift towards circular systems and integrated desalination and brine treatment technologies for resource recovery, there is a need for a unified tool. Our software addresses this need by integrating a diverse range of technologies—reverse osmosis, nanofiltration, multi-effect distillation, chemical precipitation, eutectic freeze crystallization, electrodialysis, and thermal crystallization—into a comprehensive platform. This platform not only models these processes but also provides detailed techno-economic and environmental analyses.
+
 The software provides a variety of examples to help modellers design and evaluate different technical configurations. An open-access simulation tool is crucial for enhancing the credibility, repeatability, and comparability of desalination studies, and for supporting informed design and decision-making. By offering transparent and accessible models, our software aims to advance research and practice in desalination and resource recovery.
 
 # Limitations 
