+++ example/Example_1_Tutorial.md
@@ -66,12 +66,7 @@ You can initialize the feed solution by setting the flow rate, specifying the fo
 components = ['Na', 'Cl', 'K', 'Mg', 'Ca', 'SO4']
 Ci_in = [12.33, 21.67, 0.45, 1.39, 0.45, 3.28]
 z_values = [1, -1, 1, 2, 2, -2]
-
-    # Feed flowrate
-Qsw = 3000 / 24 * d_in #kg/hr
-Qf = Qsw  # kg/hr
 ```
-Note that if you want to add more components, you need to update the components list and include the concentration of the new component in the _Ci_in_
 
 You can calculate the density of the feed solution:
 ```python
@@ -81,6 +76,15 @@ T=20+273 #Operating temperature (units: K)
     # Feed flow density 
 d_in = density_calc(T-273, mg_in)  # kg/m3
 ```
+
+```python
+    # Feed flowrate
+Qsw = 3000 / 24 * d_in #kg/hr
+Qf = Qsw  # kg/hr
+```
+> **Note:**
+> 
+> Note that if you want to add more components, you need to update the components list and include the concentration of the new component in the `Ci_in`.
 
 ## 3.2. Use process unit model
 
