+++ example/Example_1_Tutorial.md
@@ -203,7 +203,10 @@ d_in = density_calc(25, sum(Cin_mfpfr)) / 1000
 ##### Setting other input parameters
 Then the required input for MFPFR unit need to be added from unser. 
 
-First, the concentration of the alkaline solution (NaOH) and acid solution (HCl) are import. Note that different chemicals and concentrations can be used for the percicipation and the pH neutralization.
+First, the concentration of the alkaline solution (NaOH) and acid solution (HCl) are import. 
+> **Note:**
+> 
+> Note that different chemicals and concentrations can be used for the percicipation and the pH neutralization.
 
 ```python
     # Concentration of NaOH solution for step 1 and step 2 in MOL/L
@@ -321,7 +324,9 @@ Calculate the total pumping energy including the HCl stream
 E_el_mfpf=(Epump_1+Epump_2+(QHCl*dp_HCl)*1e5/3600/(1000*npump))/1000
 print("Total electricity energy consumption is "+str(round(E_el_mfpf,2))+ " KW")
 ```
-Note that you can add a calculation for filtration unit and then sum the energy requirements. 
+> **Note:**
+> 
+> Note that you can add a calculation for filtration unit and then sum the energy requirements. 
 Specific energy consumption can also be calculated: 
 ```python
     # Specific energy consumption per kg of Mg(OH)2, KWh/kg of Mg(OH)2
@@ -348,7 +353,9 @@ You need to follow similar steps for the other two processes.
 |                                           | Ion concentration [g/L]                     | Flow rate of concentrate stream [m³/h] and composition [g/L]        |
 |                                           | Current density [A/m²]                      | Electricity requirements [kWhel]                     |
 
-_Note that the feed flow rate and concentration of the units are the effluent flow rate and ions concentration of the unit before in the treatment chain._ 
+> **Note:**
+> 
+> Note that the feed flow rate and concentration of the units are the effluent flow rate and ions concentration of the unit before in the treatment chain. 
 In this treatment chain, Electrodialysis with bipolar membrane has two streams as feed for the salt channel. The two streams are mixed. For this the following calculations are required to calculate the new flow rate and concentration after the mixing. 
 ```python
     # Feed flow rate L/h
@@ -479,7 +486,9 @@ for i in range(len(prd)):
     reve_t = reve_t+rev_calc.rev_prd
     reve_list.append(rev_calc.rev_prd)
 ```
-**_Note that a detailed description of the economic model and more economic indicators can be found in ._** 
+> **Note:**
+> 
+> Note that a detailed description of the economic model and more economic indicators can be found in [Economic tutorial](https://github.com/rodoulak/desalsim/blob/main/Tutorials/Economic_Tutorial.md).  
 
 ### 4.2.3. Environmental indicators 
 A simple indicator to evaluate the environmental performance of the treatment chain is the **Carbon dioxide emissions**.  
