### 2024-09-11 5d232bd
Update MED_Tutorial.md
