+++ Tutorials/Example_1_Tutorial.md
@@ -16,7 +16,7 @@ Each section guides you through setting up the simulation environment, running t
 
 **Figure 1**. Process flow diagram of example 1.
 <br>
-
+</div>
 ## 2. Installation 
 The easiest way is through pip, in command-line interface:   
 ```
@@ -196,6 +196,11 @@ Power for pump (KW): 60.01
 E_el_nf (KW): 75.02  
 Specific Energy Consumption (KWh/m3 of permeate): 0.85
 
+##### Chemical consumption 
+```python
+QHCl_nf=0
+Qantsc_nf=0
+```
 
 ### 3.2.2.  Multi-plug flow reactor
 
@@ -400,48 +405,105 @@ After the simulation of the treatment chain, the performance of the process unit
 The following code can be used to summarise the most important results from each process unit. Note that more results can be added. 
 ```python
 class indic:
-    
-    def __init__(self, tech, Qout, Qin, Qprod_1, prod_1_name, Qprod_2, prod_2_name, E_el, E_th, Cin, Cout, chem1,chem2):
-        self.tech=tech
-        self.Qout=Qout
-        self.Qin=Qin
-        self.E_el=E_el
-        self.E_th=E_th
-        self.Cin=Cin
-        self.Cout=Cout
-        self.Qprod_1=Qprod_1
-        self.Qprod_2=Qprod_2
-        self.prod_1_name=prod_1_name
-        self.prod_2_name=prod_2_name
-        self.chem=chem1+chem2
+    """
+    A class to summarize the performance indicators of a given technology.
+
+    Attributes
+    ----------
+    tech : str
+        The name of the technology (e.g., "NF" for Nanofiltration).
+    Qout : float
+        The output flow rate (e.g., m³/h).
+    Qin : float
+        The input flow rate (e.g., m³/h).
+    Qprod_1 : float
+        The flow rate of the first product stream (e.g., m³/h).
+    prod_1_name : str
+        The name of the first product stream (e.g., "water").
+    Qprod_2 : float
+        The flow rate of the second product stream, if any (e.g., m³/h).
+    prod_2_name : str
+        The name of the second product stream, if any.
+    E_el : float
+        The electrical energy consumption (e.g., kWh).
+    E_th : float
+        The thermal energy consumption (e.g., kWh).
+    Cin : list
+        The concentration of components in the input stream.
+    Cout : list
+        The concentration of components in the output stream.
+    chem : float
+        The total chemical consumption (sum of chem1 and chem2).
+    """
+
+    def __init__(self, tech, Qout, Qin, Qprod_1, prod_1_name, Qprod_2, prod_2_name, E_el, E_th, Cin, Cout, chem1, chem2):
+        self.tech = tech
+        self.Qout = Qout
+        self.Qin = Qin
+        self.E_el = E_el
+        self.E_th = E_th
+        self.Cin = Cin
+        self.Cout = Cout
+        self.Qprod_1 = Qprod_1
+        self.Qprod_2 = Qprod_2
+        self.prod_1_name = prod_1_name
+        self.prod_2_name = prod_2_name
+        self.chem = chem1 + chem2
+
+    def techn_indi(self):
+        """
+        Calculate and return the technical indicators for the technology.
         
+        This method calculates the water recovery rate (if applicable) and
+        assigns it to the instance. It also handles cases where no water is produced.
         
-    def techn_indi(self):        
-        if self.prod_1_name=="water": 
-            self.Water_rr=self.Qprod_1/self.Qin*100 # %
-            self.Qwater=self.Qprod_1
-        elif self.prod_2_name=="water":
[…]
