+++ Tutorials/Economic_Tutorial.md
@@ -48,7 +48,7 @@ inf=0.02 #(-)
 Import 'constants.py' for more constant parameters like molecular weights (MW) etc.  
 
 ```python
-from desalsim import constants
+import desalsim.constants
 ```
 
 ### 2.3. Set input data 
