+++ Tutorials/Example_1_Tutorial.md
@@ -66,8 +66,7 @@ import desalsim.scaleup
 ```
 For example: 
 ```python
-#Molecular weight 
-    #molecular weight
+# Molecular weights 
 MW_Na=constants.MW_Na
 MW_Cl=constants.MW_cl
 MW_SO4=constants.MW_so4
