+++ Desalsim/edbm_unit_f.py
@@ -1,8 +1,9 @@
 import math 
 import numpy as np
-from desalsim import scaleup
-from desalsim.density_calc import density_calc 
-from desalsim import constants 
+from . import scaleup
+from .density_calc import density_calc 
+from . import constants
+
 
 #%%
 #Molecular weight 
