+++ Desalsim/mfpfr_unit_f.py
@@ -1,8 +1,8 @@
 import math
-from desalsim import scaleup
-from desalsim.density_calc import density_calc 
+from . import scaleup
+from .density_calc import density_calc 
 import math
-from desalsim import constants 
+from . import constants 
 #%%
     #molecular weight
 MW_Na=constants.MW_Na
