+++ Desalsim/nanofiltration_unit_f.py
@@ -1,7 +1,7 @@
 import math
 import numpy as np
-from .density_calc import density_calc
-from . import constants 
+from desalsim.density_calc import density_calc
+from desalsim import constants 
 import math
 
 #
