+++ Desalsim/ed_unit_f.py
@@ -1,8 +1,8 @@
 import numpy as np
-from . import scaleup
-from . import constants 
+from desalsim import scaleup
+from desalsim import constants 
 import math
-from .density_calc import density_calc 
+from desalsim.density_calc import density_calc 
 
 #%% Calculations 
 """
