+++ Desalsim/thermal_cryst_f.py
@@ -1,7 +1,7 @@
 import math
-from .density_calc import density_calc 
-from . import constants 
-from . import scaleup
+from desalsim.density_calc import density_calc 
+from desalsim import constants 
+from desalsim import scaleup
 
 #%%
     #Molecular weight 
