+++ desalsim.egg-info/requires.txt
@@ -1,0 +1,7 @@
+numpy
+pandas
+scipy
+matplotlib
+scikit-learn
+plotly
+openpyxl
