+++ Desalsim/efc_unit_f.py
@@ -429,7 +429,11 @@ def CrystalSizeDistribution():
     #Calculation of the equilibrium values using interpolation/extrapolation of the solubility and ice lines
     C_compound1_eq=interpolate.splev(Var.Treactor, tck, der=0)
     
-    C_compound1_eq=float(C_compound1_eq)
+    if np.isscalar(C_compound1_eq):
+        C_compound1_eq = float(C_compound1_eq)
+    else:
+        C_compound1_eq = C_compound1_eq.astype(float)
+
     print("C_compound1_eq is "+str(C_compound1_eq))
           
     T_ice_eq_array=lm.predict(Var.C_compound1)
@@ -466,7 +470,8 @@ def CrystalSizeDistribution():
     Ng_compound1_cr=-(G_compound1_cr*Var.N_compound1_cr[0]*deltat)/deltal
     Nd_compound1_cr=-(D_compound1_cr*Var.N_compound1_cr[0]*deltat)/deltal
    
-    Var.N_compound1_cr[0]=Var.N_compound1_cr[0]+Nb_compound1_cr+Ng_compound1_cr+Nd_compound1_cr-(Fout_compound1_cr*deltat*Var.N_compound1_cr[0])/Var.M_compound1_cr
+    Var.N_compound1_cr[0] += Nb_compound1_cr + Ng_compound1_cr + Nd_compound1_cr - (Fout_compound1_cr * deltat * Var.N_compound1_cr[0]) / Var.M_compound1_cr
+
     Var.V_compound1_cr_distr[0]=Var.N_compound1_cr[0]*Compound1.kv*0.25*(deltal**4)/deltal
     atot_compound1_cr=atot_compound1_cr+Var.N_compound1_cr[0]*Compound1.ka*(deltal**2)*(0**2)
     
@@ -475,7 +480,7 @@ def CrystalSizeDistribution():
         Ng_compound1_cr=(G_compound1_cr*deltat*(Var.N_compound1_cr[i-1]-Var.N_compound1_cr[i]))/deltal
         Nd_compound1_cr=(D_compound1_cr*deltat*(Var.N_compound1_cr[i+1]-Var.N_compound1_cr[i]))/deltal
 
-        Var.N_compound1_cr[i]=Var.N_compound1_cr[i]+Ng_compound1_cr+Nd_compound1_cr-(Var.N_compound1_cr[i]*Fout_compound1_cr*deltat)/Var.M_compound1_cr
+        Var.N_compound1_cr[i] += Ng_compound1_cr + Nd_compound1_cr - (Var.N_compound1_cr[i] * Fout_compound1_cr * deltat) / Var.M_compound1_cr
         Var.V_compound1_cr_distr[i]=Var.N_compound1_cr[i]*Compound1.kv*0.25*((((i+1)*deltal)**4)-((i*deltal)**4))/deltal
         atot_compound1_cr=atot_compound1_cr+Var.N_compound1_cr[i]*Compound1.ka*(deltal**2)*(i**2)     
         
@@ -483,7 +488,7 @@ def CrystalSizeDistribution():
     Ng_compound1_cr=(G_compound1_cr*deltat*Var.N_compound1_cr[classes-2])/deltal
     Nd_compound1_cr=-(D_compound1_cr*deltat*Var.N_compound1_cr[classes-1])/deltal
     
[…]
