### 2024-09-11 #7062 @rodoulak
@rkingsbury, thank you for your comments. @elbeejay, I addressed all the comments from @rkingsbury on the paper.
