### 2024-09-11 #7062 @rodoulak
@BenWinchester, thank you for your comments. @elbeejay, I addressed all the comments except the one on the '__init__.py' file.
Regarding the __init__.py file, please note that the package is named desalsim and should be imported using import desalsim, rather than import Desalsim.
I’m happy to discuss any further questions or concerns.
