### 2024-09-11 #7062 @BenWinchester
@rodoulak , thanks for addressing, will take a look through again and check off the ones that are done :smiley:

For `__init__.py`, I'm not sure what you mean by
> should be imported using desalsim rather than import Desalsim

Is your `__init__.py` file to expose APIs that are to be accessed externally, _i.e._, when the package is installed? Currently, it doesn't seem to be being used as such: you can import modules that aren't explicitly stated in `__init__.py` when you install the package which is available on `pip` and, if you build the `.whl` (wheel) from source and install, the package fails during the `__init__.py` file because `__init__.py` appears to be wanting to import from the installed package rather than exposing the methods/classes/modules from the repository.

I hope the above makes sense? You can take a look at the [documentation](https://docs.python.org/3/reference/import.html#submodules) for the `__init__.py` files to see why this error is occurring.

(Will take a look through the examples to check that they now run :smiley: )
