### 2024-09-12 #7062 @BenWinchester
> @BenWinchester, thank you for your comments. @elbeejay, I addressed all the comments except the one on the `__init__.py` file. Regarding the `__init__.py` file, please note that the package is named desalsim and should be imported using import desalsim, rather than import Desalsim. I’m happy to discuss any further questions or concerns.

@rodoulak , I started going through the examples again and they're failing this time for a different reason. _E.G._, in [Example 1](), the code imports `desalsim.constants`,
```python
from desalsim.density_calc import density_calc
import desalsim.constants
import desalsim.scaleup
```
and then tries to access variables from the constants module
```python
# Molecular weights
MW_Na=constants.MW_Na
MW_Cl=constants.MW_cl
MW_SO4=constants.MW_so4
MW_K=constants.MW_K
MW_Ca=constants.MW_Ca
MW_Mg=constants.MW_Mg
MW = [MW_Na, MW_Cl, MW_K, MW_Mg, MW_Ca, MW_SO4]
```
However, the issue is that the import doesn't utilise the `as` functionality within Python:
```python
>>> import desalsim.constants
>>> [entry for entry in locals() if "constants" in entry]
[]
>>> # There is no entry called `constants` available in the namespace
>>> import desalsim.constants  as constants
>>> [entry for entry in locals() if "constants" in entry]
['constants']
>>> # Using `as` in the import function call allows for the code to work
```

Please go through both example markdown files and correct all issues, _i.e._, check that the instructions work from start-to-end 😄 (I'm happy to keep checking for errors like this, but I think it will cause a lot of churn on this review thread!)

Please let me know if anything is unclear 😄
