### 2024-09-12 #7062 @rodoulak
Hi @BenWinchester,

Thank you very much for taking the time to go through the examples again.

Regarding the first point about the `constants` module and the `scaleup` (see There are also now errors occurring (typos) within the latest (1.0.9.1) update:), I have updated the imports  to:
```python
from desalsim import constants
```
and
```python
from desalsim import scaleup
```
which should resolve the issue with accessing the variables. I’ve tested these changes, and they are now functioning correctly.

For the second point concerning the variable Mc in section 3.2.2, I realize that this may have caused some confusion. The example tutorial is designed to demonstrate how users can integrate different technologies and analyze results, but it does not walk through every step in detail for all the technologies included in Example 1. Specifically, it focuses on the two first units (nanofiltration and MF-PFR) and explains the general approach to calculating the concentration of a mixed stream. The complete steps for the other two technologies in Example 1 are detailed in their respective tutorials. However, I understand that this might not have been entirely clear, so I will make it more explicit in the documentation to avoid any further confusion.

Additionally, I reviewed both example markdown files to ensure all instructions work seamlessly from start to finish and correct any errors or typos that may have occurred in the latest update.

Please don’t hesitate to reach out if you have any further questions or suggestions 😄.
