### 2024-09-12 #7062 @BenWinchester
@rodoulak , thanks for getting back so quickly! 😄 I'll take a look through the examples and see whether they're all working.

For the bug in `example_1.py` for version `1.0.9.1`, did you push a pypi fix btw?

There's also the [comment re: functionality documentation](https://github.com/openjournals/joss-reviews/issues/7062#issuecomment-2340292969) where the exposed APIs aren't documented. Let me know when you've addressed it and I'll take a look at the code again.

Also, there's still the `__init__.py` error which means that the code can't be installed from a version downloaded from Github (as is suggested in the documentation) as the `.whl` file that builds has errors.

Let me know when you've fixed the above and I'll take a look through again: it's just the installation, functionality and functionality-documentation checkboxes to cross out [above](https://github.com/openjournals/joss-reviews/issues/7062#issuecomment-2260302876) 😃
