### 2024-09-12 #7062 @rkingsbury
> @rkingsbury, thank you for your comments. @elbeejay, I addressed all the comments from @rkingsbury on the paper.

Thank you @rodoulak . I'll take a look.
