### 2024-09-12 #7062 @rkingsbury
Hi @rodoulak , see additional comments regarding the references in the revised pdf:

> The list of references is complete. However, there appear to be many references in the Ref List which are not cited in the paper. Some of these appear to be cited in the Mathematical description .pdf, but that file has its own reference list.

In the revised .pdf I still see only 9 citations in the paper but 19 references in the reference list. Unless the journal policy is that all SI reference should _also_ be included in the main text, anything that is not cited in the paper should be removed. (Maybe @elbeejay can comment?)

Also can you include a DOI for the authorea preprint reference?

> Related - Thanks to @yalinli2 's comments I see that there is an additional pdf called Mathematical Description, but I do not see this pdf referenced anywhere in the main paper or in the documentation. What is it's intended purpose?

Thank you for adding an explicit reference to the Mathematical Description. To minimize ambiguity, can you clean up the link in the revised pdf to use the same term for the file and the supporting information? E.g., just link the text "Supporting Information" and rename that file to "Supporting Information", or revise the text to say "are provided in the Mathematical Description"?

![image](https://github.com/user-attachments/assets/17ef2dc7-7c41-43da-b29f-cdda5e9271a1)
