### 2024-09-12 #7062 @rkingsbury
> Related, line 36: I feel your statement of need should mention the U.S. Dept. of Energy's [WaterTAP](https://watertap.readthedocs.io/en/stable/), which is a significant, recent, open-access effort to develop a process simulation platform for water desalination and brine treatment. I think your contribution is valuable and distinct, but it would be helpful to the "average" reader for you to briefly (1 sentence) explain why they might choose to use DesalSim and not WaterTAP. (For example, different model coverage, simpler to learn, etc.)

@rodoulak thank you for including a reference to WaterTAP.  To more clearly differentiate your software's capabilities, please revise this sentence slightly (because WaterTAP does include [unit models](https://watertap.readthedocs.io/en/stable/technical_reference/unit_models/index.html) for RO, NF, and some crystallization)

> While it provides valuable simulation capabilities, it mainly focuses on desalination technologies and lacks ~~extensive~~ several important brine treatment technologies such as chemical precipitation and crystallization.

In addition, either in the sentence above or in the section below, I suggest you add a statement that `desalsim` is perhaps intended for a slightly different audience. WaterTAP is very sophisticated and powerful but has a steep learning curve and requires considerable  knowledge of python and numerical solvers. `desalsim`, on the other hand, appears much simpler for a non-expert researcher to set up and run.

> there is a need for a unified tool. Our software addresses this need by integrating a diverse range of technologies—reverse osmosis, nanofiltration, multi-effect distillation, chemical precipitation, eutectic freeze crystallization, electrodialysis, and thermal crystallization—into a comprehensive platform. This platform not only models these processes but also provides detailed techno-economic and environmental analyses.

I am only placing such emphasis on this aspect because I can tell you, at least in the U.S., that WaterTAP has gotten a lot of attention in the desalination community in recent years, and I think in order for `desalsim` to reach the audience and have the impact that you intend, it is important in your paper to be as clear as you can about what sets it apart / why someone might want to use it instead of WaterTAP (since at first glance, they appear to be trying to achieve similar goals).

Finally, in the reference the authors are listed as "contributors, W." which is a bit awkward. Can you change to perhaps U.S. Dept. of Energy?
