+++ paper/paper.md
@@ -57,7 +57,7 @@ Our software offers a comprehensive modelling framework for simulating desalinat
 Through our software's simulations, researchers, engineers, and policymakers gain the power to evaluate the resource recovery potential, economics, and greenhouse gas emissions associated with different configuration combinations. This empowerment with crucial information for early-stage design and strategic planning is a significant step toward fostering more sustainable water management practices.
 
 # Statement of need
-Traditionally, simulation models were developed to evaluate the influence of certain parameters on the characteristics of the recovered products and the performance of the technology in terms of energy, chemicals, and water consumption. However, in the desalination field, open-access simulation tools are notably lacking. While commercial software programs, like WAVE [@dupontwebsite], exist for membranes, and numerous publications discuss techno-economic models for desalination [@el2002fundamentals] and brine treatment technologies [@xevgenos2015design, @micari2020techno, @chen2021zero, @panagopoulos2021beneficiation, @poirier2022techno, @morgante2022valorisation], there is a noticeable absence of open-access simulation tools in the literature. The WaterTAP platform [@watertap] offers an open-source library for modelling water treatment technologies like reverse osmosis and electrodialysis. While it provides valuable simulation capabilities, it mainly focuses on desalination technologies and lacks extensive brine treatment.
+Traditionally, simulation models were developed to evaluate the influence of certain parameters on the characteristics of the recovered products and the performance of the technology in terms of energy, chemicals, and water consumption. However, in the desalination field, open-access simulation tools are notably lacking. While commercial software programs, like WAVE [@dupontwebsite], exist for membranes, and numerous publications discuss techno-economic models for desalination [@el2002fundamentals] and brine treatment technologies [@xevgenos2015design; @micari2020techno; @chen2021zero; @panagopoulos2021beneficiation; @poirier2022techno; @morgante2022valorisation], there is a noticeable absence of open-access simulation tools in the literature. The WaterTAP platform [@watertap] offers an open-source library for modelling water treatment technologies like reverse osmosis and electrodialysis. While it provides valuable simulation capabilities, it mainly focuses on desalination technologies and lacks extensive brine treatment.
 
 With the shift towards circular systems and integrated desalination and brine treatment technologies for resource recovery, there is a need for a unified tool. Our software addresses this need by integrating a diverse range of technologies—reverse osmosis, nanofiltration, multi-effect distillation, chemical precipitation, eutectic freeze crystallization, electrodialysis, and thermal crystallization—into a comprehensive platform. This platform not only models these processes but also provides detailed techno-economic and environmental analyses.
 
@@ -69,7 +69,7 @@ The proposed software is not designed to replace detailed physical models or sys
 # Acknowledgements 
 The software was developed by Rodoula Ktori, with theoretical support from all co-authors. Technological experts conducted the validation of each simulation model for the respective technology: Nanofiltration: Dionysia Diamantidou, Niels van Linden; Multi-effect Distillation: Alessandro Trezzi; MF-PFR: Fabrizio Vassallo, Carmelo Morgante, Andrea Cipollina; EDBM: Giovanni Virruso, Andrea Culcasi; EFC: Marcos Rodriguez Pascual.
 
[…]
