### 2024-09-12 a61954f
Update paper.bib
