+++ paper/paper.bib
@@ -186,7 +186,7 @@ Micale, G.},
   publisher={Authorea}
 }
 @misc{watertap,
-  author      = {WaterTAP contributors},
+  author      = {U.S. Dept. of Energy},
   title       = {WaterTAP: An open-source water treatment model library (Version 0.6)},
   year        = {2024},
   url         = {https://github.com/watertap-org/watertap},
