+++ paper/paper.md
@@ -61,8 +61,7 @@ Traditionally, simulation models were developed to evaluate the influence of cer
 
 With the shift towards circular systems and integrated desalination and brine treatment technologies for resource recovery, there is a need for a unified tool. Our software, Desalsim addresses this need by integrating a diverse range of technologies—reverse osmosis, nanofiltration, multi-effect distillation, chemical precipitation, eutectic freeze crystallization, electrodialysis, and thermal crystallization—into a comprehensive platform. This platform not only models these processes but also provides detailed techno-economic and environmental analyses.
 
-Though WaterTAP is powerful, it requires considerable expertise in Python programming and numerical methods. DesalSim is designed to be more accessible, making it easier for researchers without advanced programming skills to perform simulations and analyze results. This ease of use makes Desalsim particularly valuable for those exploring desalination and brine treatment technologies, especially when detailed techno-economic and environmental assessments are needed. 
-The software provides a variety of examples to help modellers design and evaluate different technical configurations. 
+Though WaterTAP is powerful, it requires considerable expertise in Python programming and numerical methods. Desalsim is designed to be more accessible, making it easier for researchers without advanced programming skills to perform simulations and analyze results. This ease of use makes Desalsim particularly valuable for researchers exploring desalination and brine treatment technologies. It is especially useful when detailed techno-economic and environmental assessments are required. The software provides a variety of examples to help modellers design and evaluate different technical configurations. 
 
 By offering transparent and accessible models, DesalSim aims to enhance the credibility, repeatability, and comparability of desalination studies, supporting informed design and decision-making in the field of desalination and resource recovery.
 
