### 2024-09-12 #7062 @rodoulak
Hi @rkingsbury,

Thank you again for your comments!

 I have revised the paper accordingly:

1. **Reference List Issue:** I reviewed the reference list and confirmed that all references are cited in the main text. However, there seems to be an issue when converting the file to PDF. For example:

_"However, in the desalination field, open-access simulation tools are notably lacking. While commercial software programs, like WAVE [@dupontwebsite], exist for membranes, and numerous publications discuss techno-economic models for desalination [@el2002fundamentals] and brine treatment technologies [@xevgenos2015design; @micari2020techno; @chen2021zero; @panagopoulos2021beneficiation; @poirier2022techno; @morgante2022valorisation], there is a noticeable absence of open-access simulation tools in the literature. "_.
I replaced the commas with semicolons to separate references, but I’m not certain if this will resolve the issue in the PDF. If you have any suggestions for ensuring correct reference formatting, I would appreciate your guidance.

2. **Supplementary Information:** To avoid any confusion, I have renamed the _"Supplementary Information"_ to _"Mathematical Description"_ document.
4. **Software Capabilities:** I revised the manuscript text to reflect your suggestions regarding the software’s capabilities and differentiation from other tools like WaterTAP.
5. **WaterTAP Citation:** The author attribution for WaterTAP has been updated as per your recommendation.
