### 2024-09-12 #7062 @elbeejay
> Hi @rodoulak , see additional comments regarding the references in the revised pdf:
>
> > The list of references is complete. However, there appear to be many references in the Ref List which are not cited in the paper. Some of these appear to be cited in the Mathematical description .pdf, but that file has its own reference list.
>
> In the revised .pdf I still see only 9 citations in the paper but 19 references in the reference list. Unless the journal policy is that all SI reference should _also_ be included in the main text, anything that is not cited in the paper should be removed. (Maybe @elbeejay can comment?)
>

Thanks for flagging this @rkingsbury, you are correct, we only want references in the paper itself to be included in the paper's reference list. I agree with the comment regarding the proposed change to the "supporting information" text as well.
