### 2024-09-13 #7062 @rodoulak
Hi @BenWinchester,

Thank you for your detailed feedback. I’ve reviewed the issues related to the `__init__.py` file and made the following updates:

 **1. Relative Imports:** I’ve replaced the absolute imports with relative imports in the __init__.py file (according to the [documentation](https://docs.python.org/3/reference/import.html#submodules)) to ensure that the modules are correctly imported from the local repository, even when the package is installed directly from GitHub.

**2. Installation Testing:** I’ve tested the package installation from source and verified that it works correctly without errors.

**3. Clarification on the Purpose of `__init__.py` ([see comment](https://github.com/openjournals/joss-reviews/issues/7062#issuecomment-2344138462)):** The primary purpose of the `__init__.py` file in our package is indeed to expose key functions and classes that users will need to access when they import the package. We have structured the file to ensure that these APIs are readily available and easy to use.

**4. Regarding typo errors in the latest version ( [see comment](https://github.com/openjournals/joss-reviews/issues/7062#:~:text=in%20the%20scope.-,There%20are%20also%20now%20errors%20occurring%20(typos)%20within%20the%20latest%20(1.0.9.1)%20update%3A,-%3E%3E%3E%20import)):** The errors have been resolves, the scaleup function has been updated in the examples (see [line 232 in `example_1.py`](https://github.com/rodoulak/desalsim/blob/97a0f3b1dfab2ba9c0066b0cf6046b8c60d32bc4/example/example_1.py#L232),[line 290 in `example_1.py`](https://github.com/rodoulak/desalsim/blob/97a0f3b1dfab2ba9c0066b0cf6046b8c60d32bc4/example/example_1.py#L290), and [line 274 in `example_1.py`](https://github.com/rodoulak/desalsim/blob/97a0f3b1dfab2ba9c0066b0cf6046b8c60d32bc4/example/example_1.py#L274).

Please let me know if there are any further issues with the new `__init__.py` or if you have any further questions or comments.
