### 2024-09-24 #7062 @BenWinchester
@rodoulak , thanks for addressing the comments. Just a few things still need fixing on my end:

- [ ] For the functionality documentation:
  - `density_calc`, `HClAddition`, `energycons`, `ElectrodialysisCalc`, and `econom` are still missing docstrings. For `density_calc`, for the docstring to be properly accessible when the API is used, the docstring should be a functional dosctring rather than a module docstring;
- [ ] As per [comments above](https://github.com/openjournals/joss-reviews/issues/7062#issuecomment-2345751338), the examples provided still aren't working. Currently, [3.2.2.](https://github.com/openjournals/joss-reviews/issues/7062#issuecomment-2345751338) is failling as `Mc` isn't defined. Please go through the examples you provide fully and ensure that there are no bugs and issues 🙂
