+++ Tutorials/Example_1_Tutorial.md
@@ -371,8 +371,8 @@ print("Specific energy consumption per brine intake is "+str(round(SEC_el_feed,2
 Specific energy consumption per product is 2.88 KWh/kg product  
 Specific energy consumption per brine intake is 1.49 KWh/m3 of feed 
 
-### 3.2.2.  Other units
-You need to follow similar steps for the other two processes. 
+### 3.2.3.  Other units
+You need to follow similar steps like Sections [3.2.1.](#3.2.1.Nanofiltration) and [3.2.2.](#3.2.2.Multi-plug_flow_reactor) for the other two processes. 
 **Table 3** gives an overview of the main inputs and outputs for each process unit of Electrodialysis with bipolar membranes and Electrodialysis. 
 | Process                                   | Input                                       | Output                                                |
 |-------------------------------------------|---------------------------------------------|-------------------------------------------------------|
@@ -385,6 +385,9 @@ You need to follow similar steps for the other two processes.
 |                                           | Current density [A/m²]                      | Electricity requirements [kWhel]                     |
 
 > [!NOTE]
+> For Electrodialysis with bipolar membrane can find detailed steps in the [EDBM Tutorial](Tutorials/EDBM_Tutorial.md). 
+
+> [!IMPORTANT]
 > Note that the feed flow rate and concentration of the units are the effluent flow rate and ions concentration of the unit before in the treatment chain. 
 In this treatment chain, Electrodialysis with bipolar membrane has two streams as feed for the salt channel. The two streams are mixed. For this the following calculations are required to calculate the new flow rate and concentration after the mixing. 
 ```python
