### 2024-09-24 5a1045d
create ED_Tutorial
