+++ Tutorials/ED_Tutorial.md
@@ -1,0 +1,314 @@
+# Tutotrial: Electrodialysis with Bipolar membranes (EDBM) unit
+
+EDBM is a membrane-based technology that allows the production of acidic and alkaline solutions by applying an electric potential to the electrodes. In this work, EDBM aims to convert NaCl molecules of a brine solution to NaOH and HCl solutions. 
+![edbm](https://github.com/rodoulak/Desalination-and-Brine-Treatment-Simulation-/assets/150446818/2580e05b-73f2-4fa4-8d52-71762088fc17)
+
+
+In **desalsim** package, the EDBM unit is used to model the operation of a Electrodialysis with Bipolar membranestechnology. Upon simulation,  it calculates the flow rate of the acid, base and salt solutions, their ion concentration, and the electricity requirements of the unit.
+The EDBM function consists of one class: [EDBMCalc](#use-edbmcalc-class) that constsis.  
+In this tutorial, we will focus on how to use the the class and their methods. 
+
+**Table 1** gives an overview of the main inputs and outputs for each process unit of Electrodialysis with Bipolar membranes. 
+| Process                                   | Input                                       | Output                                                |
+|-------------------------------------------|---------------------------------------------|-------------------------------------------------------|
+| Electrodialysis with bipolar membranes| Feed flow rate [m3/h]                        | Flow rate of acid [m3/h] and composition [g/L]             |
+|                                        | Ion concentration [g/L]                      | Flow rate of base [m3/h] and composition [g/L]             |
+|                                        | Electric density                            | Flow rate of salt [m3/h] and composition [g/L]             |
+|                                        |                                            | Electricity requirements [kWhel]                          |
+
+The mathematical description of Electrodialysis with Bipolar membranestechnology is given in [Mathematical description](https://github.com/rodoulak/Desalination-and-Brine-Treatment-Simulation-/tree/main/paper/Mathematical_description.pdf), see Section A.6. 
+
+## 1. Getting started 
+### 1.1. Import class
+```python
+import desalsim
+```
+Then import the class:  
+```python
+from desalsim.edbm_unit_f import EDBMCalc
+```
+Additionally, function for calculating density (`density_calc.py`) or constants ('comparison.py') where user can add constant values like MW, prices etc, need to be imported.
+```python
+from desalsim.density_calc import density_calc 
+from desalsim import constants
+from desalsim import scaleup
+```
+### 1.2. Define feed characteristics
+You can initialize the feed solution by setting the flow rate, specifying the focus components and their concentration. 
+```python
+# Input conditions
+ph_s=4.71 #pH salt channel (units: -)
+ph_b=7#pH base channel (units: -)
+ph_a=7#pH acid channel (units: -)
+
+  # Feed concentration
[…]
