+++ Tutorials/ED_Tutorial.md
@@ -1,22 +1,21 @@
-# Tutotrial: Electrodialysis with Bipolar membranes (EDBM) unit
+# Tutotrial: Electrodialysis (ED) unit
 
-EDBM is a membrane-based technology that allows the production of acidic and alkaline solutions by applying an electric potential to the electrodes. In this work, EDBM aims to convert NaCl molecules of a brine solution to NaOH and HCl solutions. 
-![edbm](https://github.com/rodoulak/Desalination-and-Brine-Treatment-Simulation-/assets/150446818/2580e05b-73f2-4fa4-8d52-71762088fc17)
+ED  is a technique that uses ion exchange membranes and electricity to extract ionic compounds from aqueous solutions, with an electrical field serving as the driving force for solute separation. In this work, ED aims to purify NF permeate. 
+![ed](https://github.com/user-attachments/assets/386c055d-9935-41fa-b15c-e790fc169f9a)
 
 
-In **desalsim** package, the EDBM unit is used to model the operation of a Electrodialysis with Bipolar membranestechnology. Upon simulation,  it calculates the flow rate of the acid, base and salt solutions, their ion concentration, and the electricity requirements of the unit.
-The EDBM function consists of one class: [EDBMCalc](#use-edbmcalc-class) that constsis.  
+In **desalsim** package, the ED unit is used to model the operation of a Electrodialysis technology. Upon simulation, it calculates the flow rate of the concentrate and dilute streams, their ion concentration, and the electricity requirements of the unit.
+The ED function consists of one class: [ElectrodialysisCalc](#use-electrodialysiscalc-class ) that constsis of six methods: `Ts_cp`, `w_cp`, `Ls_cpv`, `Lw_cp`, `p_osmo`, `dC` [see Section 2.3](#23-use-ts_cp-w_cp-ls_cpv-lw_cp-p_osmo-dc-methods).  
 In this tutorial, we will focus on how to use the the class and their methods. 
 
-**Table 1** gives an overview of the main inputs and outputs for each process unit of Electrodialysis with Bipolar membranes. 
+**Table 1** gives an overview of the main inputs and outputs for each process unit of Electrodialysis. 
 | Process                                   | Input                                       | Output                                                |
 |-------------------------------------------|---------------------------------------------|-------------------------------------------------------|
-| Electrodialysis with bipolar membranes| Feed flow rate [m3/h]                        | Flow rate of acid [m3/h] and composition [g/L]             |
-|                                        | Ion concentration [g/L]                      | Flow rate of base [m3/h] and composition [g/L]             |
-|                                        | Electric density                            | Flow rate of salt [m3/h] and composition [g/L]             |
-|                                        |                                            | Electricity requirements [kWhel]                          |
[…]
