### 2024-09-24 #7062 @elbeejay
> rodoulak

> Hi @elbeejay,
>
> I finished with revisions and addressed reviewer comments.
>
> I have only one question regarding the references, as I also mentioned in a [comment above](https://github.com/openjournals/joss-reviews/issues/7062#issuecomment-2346716898): **Reference List Issue:** I reviewed the reference list and confirmed that all references are cited in the main text. However, there seems to be an issue when converting the file to PDF. For example: "However, in the desalination field, open-access simulation tools are notably lacking. While commercial software programs, like WAVE [@dupontwebsite], exist for membranes, and numerous publications discuss techno-economic models for desalination [@el2002fundamentals] and brine treatment technologies [@xevgenos2015design; @micari2020techno; @chen2021zero; @panagopoulos2021beneficiation; @poirier2022techno; @morgante2022valorisation], there is a noticeable absence of open-access simulation tools in the literature. "
>
> I replaced the commas with semicolons to separate references, but I’m not certain if this will resolve the issue in the PDF. If you have any suggestions for ensuring correct reference formatting, I would appreciate your guidance.

To group multiple citations the semi-colon should work just fine.
