### 2024-09-25 #7062 @rodoulak
Ηι @BenWinchester,

Thank you again for your feedback. I have made several updates to address your comments:
>
> * [x]  For the functionality documentation:
>
>   * `density_calc`, `HClAddition`, `energycons`, `ElectrodialysisCalc`, and `econom` are still missing docstrings. For `density_calc`, for the docstring to be properly accessible when the API is used, the docstring should be a functional dosctring rather than a module docstring;

Docstrings have been added to `density_calc`, `HClAddition`, `energycons`, `ElectrodialysisCalc`, and `econom`, ensuring clarity on functionality. The density_calc docstring is now a proper functional docstring as requested.

> * [x]  As per [comments above](https://github.com/openjournals/joss-reviews/issues/7062#issuecomment-2345751338), the examples provided still aren't working. Currently, [3.2.2.](https://github.com/openjournals/joss-reviews/issues/7062#issuecomment-2345751338) is failling as `Mc` isn't defined. Please go through the examples you provide fully and ensure that there are no bugs and issues 🙂

I revised [3.2.3 Other units](https://github.com/rodoulak/desalsim/blob/main/Tutorials/Example_1_Tutorial.md) to clarify that detailed steps for Nanofiltration (NF) and MF-PFR are provided, while EDBM and ED steps focus on calculating mixed stream flow rates and concentrations. Links to their respective tutorials have been added, and the issue with Mc has been clarified, directing users to the ED tutorial for further steps.

The code has been revised so the above comment is clear also in the calculation of the mixed stream:

```
    # Feed flow rate L/h
Q_in_edbm = M_mfpfr_out + Mc  # Where M_mfpfr_out is the effluent from MF-PFR, and Mc is the effluent from ED
# Mc is calculated in the ED tutorial, refer to ED Tutorial for detailed steps on calculating Mc (see https://github.com/rodoulak/desalsim/blob/main/Tutorials/ED_Tutorial.md)
```

The example code in [example_1.py](https://github.com/rodoulak/desalsim/blob/main/example/example_1.py) works properly.
