### 2024-10-07 rodoulak/desalsim#10 @rkingsbury
Use GitHub tags and releases plus semantic versioning for consistency with PyPi
The documentation gives examples of using `pip` to install a specific version, e.g.

```
pip install desalsim==1.0.1
```

But there is nothing in the GitHub repository to indicate what code state a given version corresponds to. It is relatively easy to use GitHub releases and/or repository tags to communicate to users which version corresponds to a given commit. That way, a user who initially does `pip install desalsim` can easily see how far "out of date" they are if not using the GitHub version.

See [pyEQL ](https://github.com/KingsburyLab/pyEQL) or [maggma ](https://github.com/materialsproject/maggma)as example repositories that do this.

Finally, it seems the latest releases use a 4 digit version number ,e.g. `v1.0.9.3`. While there is nothing inherently wrong with this, many python users and distribution tools expect some form of "semantic versioning" as described [here](https://semver.org), which relies on 3 digit version numbers. As a user, I don't quite understand how to interpret the fourth. I suggest avoiding the fourth digit in future releases.
