### 2024-10-07 rodoulak/desalsim#11 @rkingsbury
NF equation for osmotic pressure may be inaccurate at high salinities
In the Mathematical Description, Eq. A. 9 indicates that osmotic pressure in the Nanofiltration model is calculated according to the Van't Hoff equation:

![image](https://github.com/user-attachments/assets/66bf2f1d-3fce-4d4e-a67e-62ab42e598c1)

This equation is only valid at dilute concentrations. It's accuracy can be visualized in the plot below.

![image](https://github.com/user-attachments/assets/82b92480-39ac-4baa-ad27-b8ebb42e5fbf)

Since the stated purpose of `desalsim` is to analyze brine management technologies, and since osmotic pressure plays a major role in determining the energy requirements of NF and reverse osmosis, use of this equation in the model should be reconsidered.

At a minimum, osmotic pressure should be computed using the Gibbs equation based on water activity, as described [here](https://pyeql.readthedocs.io/en/latest/class_solution.html#pyEQL.Solution.osmotic_pressure). The water activity could be approximated as the mole fraction of water if necessary; even this simplification should be much more accurate than Eq. A.9 when salt concentrations exceed that of seawater.
