### 2024-10-14 rodoulak/desalsim#12 @rkingsbury
link documentation from README or elsewhere
There is a very nice [ReadtheDocs documentation page](https://desalsim-web.readthedocs.io/en/latest/) availble for this package, but as far as I can tell, it is not linked anywhere in the README file or the tutorials. It should be linked so that it is discoverable by users.
