+++ Tutorials/Economic_Tutorial.md
@@ -3,7 +3,7 @@
 ## 1. Introduction 
 Welcome to our comprehensive tutorial on running economic models. The economic model (`economic_f.py`) can be used to perform economic assessments of the treatment train or of individual units to identify economic hotspots. The economic assessment will help researchers, engineers, and decision-makers improve the system's performance and select the most economically feasible integration of technologies. 
 
-In this tutorial, we provide step-by-step instructions on how to [use economic models](#running-economic-model) and [analyse the results](#results-evaluation) obtained using **Example 1** as case study (see **Figure 1**).  
+In this tutorial, we provide step-by-step instructions on how to [use economic models](#calculate-capital-costs-(capex)) and [analyse the results](#results-evaluation) obtained using **Example 1** as case study (see **Figure 1**).  
 
 <figure>
   <img src="https://github.com/rodoulak/Desalination-and-Brine-Treatment-Simulation-/assets/150446818/55cc6b6f-dde8-4b12-ae61-fa23665c288e" alt="Image" style="width:600px;">
@@ -235,8 +235,9 @@ The calculation of chemicals and water costs is similar to the energy cost, mult
 | Fixed charges: 5% of annual OPEX                                  |
 | Plant overhead costs: 5% of annual OPEX                           |
 
+
 > [!NOTE]
-> Note that the assumptions listed in **Table 3** can be modified to suit different case studies.>
+> Note that the assumptions listed in **Table 3** can be modified to suit different case studies.
 
 ```python
    # Set assumptions for Operating cost (OPEX)
