+++ Tutorials/Tutorial.md
@@ -121,6 +121,10 @@ For the economic analysis of a full-scale desalination plant, the equipment cost
 |                                   | Plant overhead costs: 5% of annual OPEX                           |
 
 
+> [!NOTE]
+> Note that the assumptions listed in **Table 3** can be modified to suit different case studies.
+
+
 ### 6. Treatment trains comparison 
 In comparison file, results from different treatment trains are summarised. Indicators are formulated to compare the treatment trains. 
 
