### 2024-10-24 2f069c7
class docstring
