+++ desalsim.egg-info/SOURCES.txt
@@ -1,46 +1,0 @@
-LICENSE
-README.md
-setup.py
-desalsim/__init__.py
-desalsim/constants.py
-desalsim/density_calc.py
-desalsim/economic_f.py
-desalsim/ed_unit_f.py
-desalsim/edbm_unit_f.py
-desalsim/efc_unit_f.py
-desalsim/med_unit_f.py
-desalsim/mfpfr_unit_f.py
-desalsim/nanofiltration_unit_f.py
-desalsim/ro_unit_f.py
-desalsim/scaleup.py
-desalsim/thermal_cryst_f.py
-desalsim.egg-info/PKG-INFO
-desalsim.egg-info/SOURCES.txt
-desalsim.egg-info/dependency_links.txt
-desalsim.egg-info/requires.txt
-desalsim.egg-info/top_level.txt
-example/__init__.py
-example/comparison.py
-example/constants.py
-example/density_calc.py
-example/economic_f.py
-example/ed_unit_f.py
-example/edbm_unit_f.py
-example/efc_unit_f.py
-example/example_1.py
-example/example_2.py
-example/med_unit_f.py
-example/mfpfr_unit_f.py
-example/nanofiltration_unit_f.py
-example/ro_unit_f.py
-example/scaleup.py
-example/thermal_cryst_f.py
-tests/__init__.py
-tests/scaleup.py
-tests/test_economic.py
-tests/test_ed.py
-tests/test_edbm.py
-tests/test_med.py
-tests/test_mfpfr.py
-tests/test_nanofiltration.py
-tests/test_thermal_cryst.py
