+++ desalsim.egg-info/requires.txt
@@ -1,7 +1,0 @@
-numpy
-pandas
-scipy
-matplotlib
-scikit-learn
-plotly
-openpyxl
