### 2024-10-24 #7062 @rodoulak
@BenWinchester, thank you for your comments.

- [x]  I’ve updated the documentation in` economic_f.py `to include a class-level docstring for `econom,` which provides an overview of the class’s purpose and attributes. This should now properly display when the class is accessed through the API.

- [x] Regarding this [comment](https://github.com/openjournals/joss-reviews/issues/7062#:~:text=%40rodoulak%20%2C%20I%20also,this%20change%20in%3F). I want to ensure I'm addressing your concerns accurately, and I believe there might be some confusion regarding the example and the tutorial documentation.

To clarify, the example Python files already contain the complete calculations, including the calculation of Mc as part of the overall simulation. However, in the tutorial documentation, to avoid redundancy, I skipped some detailed steps and instead referred the user to the individual technology tutorials (such as for ED and EDBM) for specific instructions. My intention was to keep the tutorial concise while ensuring that the necessary information is available in the respective technology tutorials.

In particular, the ED tutorial covers the detailed steps for calculating Mc, which is linked in the tutorial [here](https://github.com/rodoulak/desalsim/blob/main/Tutorials/ED_Tutorial.md). The integration of the technologies, including mixed-stream calculations, is covered in the main tutorial.

See the updated explanation in the [tutorial for example 1](https://github.com/rodoulak/desalsim/blob/main/Tutorials/Example_1_Tutorial.md#323--other-units).
 Please let me know if you were referring to something else, or if further clarification or changes are needed.
