### 2024-10-24 #7062 @rodoulak
@rkingsbury, thank you for all the insightful comments.

@elbeejay, I made all the required revisions that were requested.

Specifically:
**1. Regarding the points in [Functionality documentation:](https://github.com/openjournals/joss-reviews/issues/7062#:~:text=scripts%20without%20issue.-,Functionality%20documentation%3A%20Is%20the%20core%20functionality%20of%20the%20software%20documented%20to%20a%20satisfactory%20level%20(e.g.%2C%20API%20method%20documentation)%3F,-The%20Tutorial.md)**
- I would like to clarify that ` desalsim` is designed as a flexible simulation tool that models different desalination and brine treatment technologies. The core functionality of ` desalsim` includes the simulation of processes such as reverse osmosis, nanofiltration, multi-effect distillation, electrodialysis, and others. These simulations include calculating flow rates, concentrations, and energy consumption for each technology, among other outputs.
However,` desalsim` relies on user input to define the parameters of each specific case study—such as concentrations, flow rates, and the specific process units that the user wishes to simulate. This allows the tool to be adaptable to a wide range of scenarios and to provide results based on the specific conditions provided by the user.
The tutorials are intended to guide users through the process of setting up their simulations, which is why they focus on scripting. For instance, the indic class in the tutorials is a means of helping users summarize and visualize their results in a customizable way. While this class is not currently part of the core functionality of ` desalsim`, it allows users the flexibility to analyze their results in a way that best suits their particular study. This flexibility gives users the freedom to extend the functionality of the tool based on their unique needs.

- "Electric density" for the ED processes has been replaced with "Current Density [A/m2]."
- economic assumptions (Table 3):  a note has been added to clarify this point:
[Note that the assumptions listed in Table 3 can be modified to suit different case studies.](https://github.com/rodoulak/desalsim/blob/main/Tutorials/Tutorial.md#:~:text=Note,different%20case%20studies.)
- The "treatment chain" has been replaced with "treatment train" in all tutorials.
- the second` import numpy as np` has been deleted in the Treatment Chains Comparison section.

**2. Regarding the [readthedocs page](https://desalsim-web.readthedocs.io/en/latest/) and [issue on it](https://github.com/rodoulak/desalsim/issues/12#issue-2586505138):**
The readthedocs page is mentioned in the readme file in the [documentation section](https://github.com/rodoulak/desalsim/tree/main#:~:text=Documentation,at%20Desalsim.).

**3. Regarding the issue on the [releases](https://github.com/rodoulak/desalsim/issues/10#issue-2571009878) :**
The [release ](https://github.com/rodoulak/desalsim/releases) has been added, and the version number has changed to 3 digit number.
