### 2024-10-25 #7062 @elbeejay
Looks good thanks @rodoulak for making revisions so quickly.

I agree with @BenWinchester about tidying up the repository, I'd agree with his suggestions to omit the `__pycache__`, `dist/` and `*.egg-info` folders and files, I'd even suggest including a `.gitignore` file to avoid committing these in the future. GitHub has a boilerplate .gitignore for Python packages: https://github.com/github/gitignore/blob/main/Python.gitignore

For the issues @rkingsbury opened up, it'd be helpful if you replied in each of the issue threads, and provided links to the commits or specific lines that were changed to make things easier on @rkingsbury when he revisits the issues. This will also help make it more clear when/why the issues get closed. Hope that makes sense, thanks!
