### 2024-10-29 rodoulak/desalsim#10 @rodoulak
Thank you for the feedback! I’ve addressed your comments as follows:

1. Release Tagging: I've added GitHub [releases](https://github.com/rodoulak/desalsim/releases) that correspond to specific code states for each version, making it easier to identify which version corresponds to a particular commit. This should now allow users to see the version history and any associated updates or fixes.
2. Versioning Format: The version numbering format has been updated to follow the standard three-digit semantic versioning (v1.1.0).

These updates should help users stay informed about the current state of their installation and make it easier to keep track of changes across versions.
