### 2024-10-29 rodoulak/desalsim#12 @rodoulak
Thank you for pointing that out! I’ve made the following updates:
ReadTheDocs Link in README: I have ensured that the link to the ReadTheDocs documentation page is prominently included in the README file, under the [documentation section](https://github.com/rodoulak/desalsim/tree/main#:~:text=Documentation,at%20Desalsim.), making it easily accessible for users visiting the GitHub repository.
