+++ example/example_2.py
@@ -100,9 +100,9 @@ Qperm = nfmass_objects[0].Qperm  # kg/hr
 Qconc= nfmass_objects[0].Qconc # kg/hr
 
 # Calculate Osmotic Pressure
-P_osmo_f = OsmoticPressure(c_values, z_values, T).osmotic_pressure_calculation()
-P_osmo_p = OsmoticPressure(Cperm, z_values, T).osmotic_pressure_calculation()
-P_osmo_c = OsmoticPressure(Cconc, z_values, T).osmotic_pressure_calculation()
+P_osmo_f = OsmoticPressure(Ci_in, MW_values).calculate_osmotic_pressure()
+P_osmo_p = OsmoticPressure(Cperm, MW_values).calculate_osmotic_pressure()
+P_osmo_c = OsmoticPressure(Cconc, MW_values).calculate_osmotic_pressure()
 
 d_p=density_calc(T-273, sum(Cperm))
 
