### 2024-10-29 rodoulak/desalsim#11 @rodoulak
Thank you for your valuable feedback. Based on your suggestion, the nanofiltration function has been revised to calculate the osmotic pressure using the Gibbs equation:
$\pi = -\left( \frac{RT}{V} \right) \ln(a_w)$
where we approximate the water activity α<sub>w </sub> as the mole fraction of water. This adjustment provides a more accurate estimation of osmotic pressure, aligning with the theoretical foundations you highlighted.

In addition:

- The package has been updated on GitHub, and a detailed changelog is provided in the [new release](https://github.com/rodoulak/desalsim/releases), explaining the revisions made to the nanofiltration function and the rationale behind them.
- Example files and tutorials ([NF tutorial](https://github.com/rodoulak/desalsim/blob/main/Tutorials/NF_Tutorial.md#3-use-osmoticpressure-class:~:text=3.%20Use%20OsmoticPressure%20class), [example 1 tutorial](https://github.com/rodoulak/desalsim/blob/main/Tutorials/Example_1_Tutorial.md#:~:text=is%2070.73g/l-,Calculate%20Osmotic%20Pressure,-For%20the%20calculation)) have been updated to reflect this new calculation approach, ensuring that users have clear guidance on using the revised function.
