### 2024-10-29 #7062 @rodoulak
@elbeejay thank you for the clarifications.

@BenWinchester, I included a `.gitignore `file to avoid committing these in the future, following @elbeejay's instructions.

@elbeejay, @rkingsbury I replied in each of the issues threads see [issue #10 ](https://github.com/rodoulak/desalsim/issues/10#issuecomment-2443465318), [issue #11](https://github.com/rodoulak/desalsim/issues/11#issuecomment-2444129277), [issue #12](https://github.com/rodoulak/desalsim/issues/12#issuecomment-2443474724)
