### 2024-12-12 #7062 @BenWinchester
Hi @elbeejay , sorry for the delay, I've had many deadlines (including reviews) in the past few weeks and this completely slipped my to-do list, apologies!

I think the repository looks good now: much tidier, with the APIs well-documented, and I'm satisfied that everything's now been actioned. I'm not clear what you meant, @rodoulak , in [your comment dated the 29th](https://github.com/openjournals/joss-reviews/issues/7062#issuecomment-2435193143) as the functionality documentation is there to ensure that all the APIs are clearly documented to someone who might use them, and it's exactly because your package works so much on user imports of methods (rather than a CLI, for instance) that it's so important, so I think we're ultimately in agreement? 😆

Sorry again for the delay!
