### 2024-12-12 #7062 @elbeejay
No worries, thank you for the follow up @BenWinchester - it's definitely a busy time of year!

@rodoulak can you take a look over the repository and paper again to make sure everything is in order. There are a few specific requests and editorial actions I'll take over the next few days to go over things from the JOSS editorial perspective as well.
