### 2024-12-16 acd7a0d
Update paper.bib
