+++ paper/paper.bib
@@ -179,11 +179,13 @@ Micale, G.},
 }
 
 @article{ktori2024designing,
-  title={Designing for the future: A Value-Sensitive Approach to Integrated Desalination and Brine Treatment},
+  title={A value-sensitive approach for integrated seawater desalination and brine treatment},
   author={Ktori, Rodoula and Parada, Mar Palmeros and Rodriguez-Pascual, Marcos and van Loosdrecht, Mark CM and Xevgenos, Dimitrios},
-  journal={Authorea Preprints},
+  journal={Sustainable Production and Consumption},
   year={2024},
-  publisher={Authorea}
+  publisher={Elsevier}, 
+  doi={10.1016/j.spc.2024.11.006},
+  url={https://doi.org/10.1016/j.spc.2024.11.006}
 }
 @misc{watertap,
   author      = {U.S. Dept. of Energy},
