### 2024-12-16 #7062 @rodoulak
@elbeejay I went through the [checklist](https://github.com/openjournals/joss-reviews/issues/7062#issuecomment-2543102782).

**Note:** I made one change in the affiliations
**Note:** the last version is v.1.1.1
**Here is the DOI:** https://doi.org/10.5281/zenodo.14499653

Some additional changes:

- [x] I have also updated one of the [references ](https://github.com/openjournals/joss-reviews/issues/7062#:~:text=treatment%20model%20lib...%0A%0A%E2%9D%8C-,MISSING%20DOIs,-%2D%2010.36227/techrxiv.170594775.55495325), from pre-print to the publication that is now online.
- [x] I have updated the link in the paper to the mathematical description paper.
