### 2024-12-17 #7062 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
