### 2024-12-18 #7062 @kthyng
@rodoulak Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "india", but please check for others.
