### 2024-12-19 19685a2
Update paper.bib
