+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{abraham2011socio,
-  title={Socio-economic \& technical assessment of photovoltaic powered membrane desalination processes for India},
+  title={Socio-economic \& technical assessment of photovoltaic powered membrane desalination processes for {India}},
   author={Abraham, Tinu and Luthra, Amit},
   journal={Desalination},
   volume={268},
@@ -23,7 +23,7 @@
   url={https://doi.org/10.1016/j.renene.2011.03.040}
 }
 @article{cassaro2023electrodialysis,
-  title={Electrodialysis with bipolar membranes for the sustainable production of chemicals from seawater brines at pilot plant scale},
+  title={Electrodialysis with Bipolar Membranes for the Sustainable Production of Chemicals from {Seawater Brines} at Pilot Plant Scale},
   author={Cassaro, Calogero and Virruso, Giovanni and Culcasi, Andrea and Cipollina, Andrea and Tamburini, Alessandro and Micale, Giorgio},
   journal={ACS Sustainable Chemistry \& Engineering},
   volume={11},
@@ -35,17 +35,18 @@
   url={https://doi.org/10.1021/acssuschemeng.2c06636}
 }
 @article{chen2021zero,
-  title={A zero liquid discharge system integrating multi-effect distillation and evaporative crystallization for desalination brine treatment},
+  title={A {Zero Liquid Discharge} System Integrating {Multi-effect Distillation} and {Evaporative Crystallization} for Desalination Brine Treatment},
   author={Chen, Qian and Burhan, Muhammad and Shahzad, Muhammad Wakil and Ybyraiymkul, Doskhan and Akhtar, Faheem Hassan and Li, Yong and Ng, Kim Choon},
   journal={Desalination},
   volume={502},
-  pages={114928},  year={2021},
+  pages={114928},  
+  year={2021},
   publisher={Elsevier}, 
   doi={10.1016/j.desal.2020.114928},
   url={https://doi.org/10.1016/j.desal.2020.114928}
 }
 @article{choi2015economic,
-  title={Economic evaluation of a hybrid desalination system combining forward and reverse osmosis},
+  title={Economic Evaluation of a Hybrid Desalination System Combining {Forward} and {Reverse Osmosis}},
   author={Choi, Yongjun and Cho, Hyeongrak and Shin, Yonghyun and Jang, Yongsun and Lee, Sangho},
   journal={Membranes},
   volume={6},
@@ -57,7 +58,7 @@
   url={https://doi.10.3390/membranes6010003}
 }
 @book{el2002fundamentals,
-  title={Fundamentals of salt water desalination},
+  title={Fundamentals of {Salt Water Desalination}},
   author={El-Dessouky, Hisham T and Ettouney, Hisham Mohamed},
   year={2002},
   publisher={Elsevier}, 
@@ -76,7 +77,7 @@
 }
 
 @article{micari2020techno,
-  title={Techno-economic analysis of integrated processes for the treatment and valorisation of neutral coal mine effluents},
+  title={Techno-economic Analysis of Integrated Processes for the Treatment and Valorisation of {Neutral Coal Mine Effluents}},
   author={Micari, Marina and Cipollina, A and Tamburini, A and Moser, M and Bertsch, Valentin and Micale, Giorgio},
   journal={Journal of cleaner production},
   volume={270},
@@ -87,7 +88,7 @@
   url={https://doi.org/10.1016/j.jclepro.2020.122472}
 }
 @article{morgante2022valorisation,
-  title={Valorisation of SWRO brines in a remote island through a circular approach: Techno-economic analysis and perspectives},
[…]
