### 2024-12-19 b2b42fb
Update paper.bib
