+++ paper/paper.bib
@@ -23,7 +23,7 @@
   url={https://doi.org/10.1016/j.renene.2011.03.040}
 }
 @article{cassaro2023electrodialysis,
-  title={Electrodialysis with Bipolar Membranes for the Sustainable Production of Chemicals from {Seawater Brines} at Pilot Plant Scale},
+  title={Electrodialysis with {Bipolar Membranes} for the {Sustainable Production} of {Chemicals} from {Seawater Brines} at {Pilot Plant Scale}},
   author={Cassaro, Calogero and Virruso, Giovanni and Culcasi, Andrea and Cipollina, Andrea and Tamburini, Alessandro and Micale, Giorgio},
   journal={ACS Sustainable Chemistry \& Engineering},
   volume={11},
@@ -35,7 +35,7 @@
   url={https://doi.org/10.1021/acssuschemeng.2c06636}
 }
 @article{chen2021zero,
-  title={A {Zero Liquid Discharge} System Integrating {Multi-effect Distillation} and {Evaporative Crystallization} for Desalination Brine Treatment},
+  title={A {Zero Liquid Discharge} System Integrating {Multi-effect Distillation} and {Evaporative Crystallization} for {Desalination Brine Treatment}},
   author={Chen, Qian and Burhan, Muhammad and Shahzad, Muhammad Wakil and Ybyraiymkul, Doskhan and Akhtar, Faheem Hassan and Li, Yong and Ng, Kim Choon},
   journal={Desalination},
   volume={502},
@@ -77,7 +77,7 @@
 }
 
 @article{micari2020techno,
-  title={Techno-economic Analysis of Integrated Processes for the Treatment and Valorisation of {Neutral Coal Mine Effluents}},
+  title={Techno-economic {Analysis} of {Integrated Processes} for the {Treatment} and {Valorisation} of {Neutral Coal Mine Effluents}},
   author={Micari, Marina and Cipollina, A and Tamburini, A and Moser, M and Bertsch, Valentin and Micale, Giorgio},
   journal={Journal of cleaner production},
   volume={270},
@@ -165,7 +165,7 @@
   url={https://doi.org/10.1016/j.desal.2024.117562}
 }
 @report{xevgenos2023deliverable,
-  title={Deliverable 3.1 Report on the design procedure including bench-scale tests for CS1 and CS2},
+  title={Deliverable 3.1 Report on the design procedure including bench-scale tests for {CS1} and {CS2}},
   author={Xevgenos, D. and Ktori, R. and van Gils, E. and Diamantidou, D. and van Linden, N. and Vassalo, F. and Morgante, C. and Culcasi, A. and Rodrigues Pascual, M. and Avramidi, M. and Trezzi, A. and Krishnan, A. and Nauta, J.T. and Cipollina, A. and 
 Micale, G.},
   year={2023},
