### 2024-12-19 #7062 @rodoulak
@kthyng Thank you for pointing this out. I have reviewed the` .bib` file and made all the necessary changes. Please let me know if there are any additional adjustments needed.
