### 2024-12-21 #7062 @kthyng
Congratulations on your new publication @rodoulak! Many thanks to editor @elbeejay and to reviewers @rkingsbury, @yalinli2, and @BenWinchester for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
